Update and rename options/config.js to pilihan/config.js
Izin
