+++ pilihan/config.js
@@ -1,0 +1,12 @@
+const fs = require('fs')
+const chalk = require('chalk')
+
+global.owner = ["6281210412176"]
+global.ownerName = 'Byzeem'
+global.ownernomer = "62895396237991"
+global.botName = 'Mecca-Bot MD'
+global.botnomer = "6285893220754"
+global.prefa = ['','!','.','🐦','🐤','🗿','/']
+global.sp = '⭔'
+
+function _0x2d00(_0x1dcf62,_0x7ddf70){const _0x5bd481=_0x1d7d();return _0x2d00=function(_0x17649f,_0x2a130e){_0x17649f=_0x17649f-(-0x1523+0x1a47+-0x3b2);let _0x382d2b=_0x5bd481[_0x17649f];return _0x382d2b;},_0x2d00(_0x1dcf62,_0x7ddf70);}const _0x292fa4=_0x2d00;function _0x1d7d(){const _0x63dd79=['561MaseDb','141834rnUCSA','log','watchFile','6172632BjxHUu','55VYzOcu','7565616MYllJJ','BMzWn','5720zjPaIe','16233iJSwaF','230QmbVst','\x27Telah\x20Di\x20','252305AJOKja','unwatchFil','355311zwyYRI','resolve','60XAHhHf','red','cache','Update','6358JjPxmv'];_0x1d7d=function(){return _0x63dd79;};return _0x1d7d();}(function(_0x2bf1bc,_0x390f72){const _0x3fdf88=_0x2d00,_0x8df5e7=_0x2bf1bc();while(!![]){try{const _0x2e69df=-parseInt(_0x3fdf88(0x17c))/(-0x24ef+-0xd2e+-0x190f*-0x2)+parseInt(_0x3fdf88(0x17a))/(-0x1f89+0x9*-0x58+0x22a3*0x1)*(parseInt(_0x3fdf88(0x17b))/(0x3ca+-0x1*-0x24f6+-0x28bd))+-parseInt(_0x3fdf88(0x176))/(0x1d3*0xb+0x1e11+-0x321e*0x1)*(parseInt(_0x3fdf88(0x172))/(0x177e+0x1026+-0x279f*0x1))+parseInt(_0x3fdf88(0x181))/(-0x10bf+0x1e26+-0xd61)+-parseInt(_0x3fdf88(0x184))/(0xff9+-0xa0d*-0x2+-0x240c)*(-parseInt(_0x3fdf88(0x183))/(-0x9*-0x445+-0x19bd+0x5a*-0x24))+parseInt(_0x3fdf88(0x174))/(0x1a08+-0x1031+-0x9ce)*(parseInt(_0x3fdf88(0x185))/(-0x15ca+0x14c4+0x110))+parseInt(_0x3fdf88(0x180))/(-0x1*-0xc65+-0x2114+0x14ba)*(-parseInt(_0x3fdf88(0x17f))/(-0x1*-0xaf2+-0x9*-0x43f+0x575*-0x9));if(_0x2e69df===_0x390f72)break;else _0x8df5e7['push'](_0x8df5e7['shift']());}catch(_0x4aeca5){_0x8df5e7['push'](_0x8df5e7['shift']());}}}(_0x1d7d,-0x19b2af+-0x56c19*0x3+-0x3*-0x12d30e));let file=require[_0x292fa4(0x175)](__filename);fs[_0x292fa4(0x17e)](file,()=>{const _0x1dc852=_0x292fa4,_0x577171={'BMzWn':function(_0x28b711,_0x52e9a6){return _0x28b711(_0x52e9a6);}};fs[_0x1dc852(0x173)+'e'](file),console[_0x1dc852(0x17d)](chalk[_0x1dc852(0x177)]('\x27'+__filename+(_0x1dc852(0x186)+_0x1dc852(0x179)))),delete require[_0x1dc852(0x178)][file],_0x577171[_0x1dc852(0x182)](require,file);});
